Use archive.org URLs available as of 2020-03-29 13:10
Earlier, we did not have a good archive.org URL for MoHFW data on
2020-03-25. As a result, archive.org URL for WoM was used as reference
but the data was picked up from MoHFW after manual observation on that
data. This caused a slight inconsistency between the data in this
project and the data in the reference URL for WoM. Now that a good
archive.org URL is available for MoHFW, this inconsistency longer
exists. As a result, this commit removes the "Exceptions" section that
was dedicated to explain the inconsistency.